requirements.txt
asgiref==3.5.2
Django==3.2.16
djangorestframework==3.14.0
greenlet==1.1.3.post0
importlib-metadata==5.0.0
peewee==3.15.3
psycopg2==2.9.5
pytz==2022.5
SQLAlchemy==1.4.42
sqlparse==0.4.3
typing_extensions==4.4.0
zipp==3.10.0
